+++ examples/connect_channels.js
@@ -15,9 +15,16 @@ deviceStatus.client = Smartglass()
 deviceStatus.client.connect('192.168.2.5', function(result){
     if(result === true){
         console.log('Xbox succesfully connected!');
-        deviceStatus.client.addManager('system_input', SystemInputChannel(0))
-        deviceStatus.client.addManager('system_media', SystemMediaChannel(1))
-        deviceStatus.client.addManager('tv_remote', TvRemoteChannel(2))
+        deviceStatus.client.addManager('system_input', SystemInputChannel())
+        deviceStatus.client.addManager('system_media', SystemMediaChannel())
+        deviceStatus.client.addManager('tv_remote', TvRemoteChannel())
+
+        // deviceStatus.client.on('_on_console_status', function(message, xbox, remote){
+        //     // console.log('CONSOLE STATE', message.packet_decoded.protected_payload)
+        //     // console.log(message.packet_decoded.protected_payload)
+        //     console.log(deviceStatus.client.getActiveApp())
+        //     //deviceStatus.client.getManager('system_input').sendCommand('nexus');
+        // });
 
         setTimeout(function(){
             console.log('Send nexus button')
@@ -55,7 +62,7 @@ deviceStatus.client.connect('192.168.2.5', function(result){
                 deviceStatus.client.getManager('system_input').sendCommand('nexus');
 
                 console.log(deviceStatus.client.getActiveApp())
-                console.log(deviceStatus.client.getManager('system_media')._media_state)
+                console.log(deviceStatus.client.getManager('system_media').getState())
             }.bind(deviceStatus), 5000)
 
             setTimeout(function(){