+++ src/channels/systemmedia.js
@@ -1,12 +1,12 @@
 var Debug = require('debug')('smartglass:channel_system_media')
 const Packer = require('../packet/packer');
 
-module.exports = function(channel_request_id)
+module.exports = function()
 {
     return {
         _channel_status: false,
         _channel_id: 0,
-        _channel_request_id: channel_request_id,
+        _channel_request_id: -1,
         _smartglass: false,
         _xbox: false,
         _media_request_id: 1,
@@ -32,7 +32,8 @@ module.exports = function(channel_request_id)
             seek: 32786, // Not implemented yet
         },
 
-        load: function(smartglass){
+        load: function(smartglass, manager_id){
+            this._channel_request_id = manager_id
             this._smartglass = smartglass
             this._smartglass.on('_on_console_status', function(message, xbox, remote, smartglass){
                 if(this._channel_status == false){
@@ -43,6 +44,7 @@ module.exports = function(channel_request_id)
                     channel_request.set('title_id', 0);
                     channel_request.set('service', Buffer.from('48a9ca24eb6d4e128c43d57469edd3cd', 'hex'));
                     channel_request.set('activity_id', 0);
+                    Debug('Send channel request on channel #'+this._channel_request_id);
 
                     xbox.get_requestnum()
                     var channel_message  = channel_request.pack(xbox)
@@ -98,6 +100,10 @@ module.exports = function(channel_request_id)
 
                 this._smartglass._send(message);
             }
+        },
+
+        getState: function(){
+            return this._media_state
         }
     }
 }